refactor: extract migrate cli module
